requirements.txt
pydub==0.25.1
click==8.1.7
numpy>=1.26.0
requests==2.31.0
python-dotenv==1.0.1
tqdm==4.67.1
yt-dlp>=2023.12.30
pandas>=2.1.4
openpyxl>=3.1.2